+++ shared/src/main/scala/latphone/Latin25Alphabet.scala
@@ -50,7 +50,8 @@ import scala.scalajs.js.annotation._
   }
   /** Ordered sequence of allowed punctuation characters.*/
   def punctuationString: String = {
-    "(),;:.?"
+    //"(),;:.?"
+    """(),;:."?!"""
   }
 
   /** Set of all recognized diphthongs.*/