+++ shared/src/test/scala/latphone/Latin24AlphabetSpec.scala
@@ -1,0 +1,11 @@
+package edu.holycross.shot.latin
+import org.scalatest.FlatSpec
+
+
+
+class Latin24AlphabetSpec extends FlatSpec {
+
+  "The Latin24Alphabet object"  should "check lots of things" in pending
+
+
+}